Move scraper to application controller
